pariwisata-recommender/backend/requirements.txt
numpy==1.24.4
Cython==0.29.36
scikit-learn==1.3.2
pandas==2.1.4
scipy==1.11.4
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic[email]==2.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
lightfm==1.17
implicit==0.7.2
scikit-surprise==1.1.3
nltk==3.8.1